requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
lxml>=4.9.0
urllib3>=1.26.0
python-dateutil>=2.8.0
pytz>=2022.1
charset-normalizer>=2.1.0
idna>=3.4

# Web scraping dependencies
selenium>=4.8.0
webdriver-manager>=3.8.0

# Optional dependencies for advanced features
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
wordcloud>=1.9.0
networkx>=2.8.0
scikit-learn>=1.1.0

# AI/ML dependencies (install separately if needed)
# openai>=0.27.0
# llama-index>=0.8.0
# transformers>=4.21.0
# torch>=1.12.0 